ADD Refactor to extract context middleware from the HTTP Server
