+++ lib/fiware-iotagent-lib.js
@@ -27,7 +27,7 @@ var request = require('request'),
     async = require('async'),
     errors = require('./errors'),
     utils = require('./asyncUtils'),
-    contextServer = require('./services/contextServer'),
+    contextServer = require('./services/northboundServer'),
     logger = require('fiware-node-logger'),
     registry,
     config;
