+++ lib/services/contextServer.js
@@ -22,22 +22,11 @@
  */
 'use strict';
 
-var http = require('http'),
-    express = require('express'),
-    async = require('async'),
-    contextServer,
+var async = require('async'),
     logger = require('fiware-node-logger'),
     updateHandler,
     queryHandler;
 
-function handleError(error, req, res, next) {
-    switch (error.name) {
-        default:
-            res.json(500, {
-                message: error
-            });
-    }
-}
 
 function createUpdateResponse(req, res, results) {
     var result = {
@@ -215,46 +204,12 @@ function handleNotification(req, res, next) {
     }
 }
 
-function traceRequest(req, res, next) {
-    logger.debug('Request for path [%s] from [%s]', req.path, req.get('host'));
-    logger.debug('Body:\n\n%s\n\n', JSON.stringify(req.body, null, 4));
-
-    next();
-}
-function start(config, callback) {
-    contextServer = {
-        server: null,
-        app: express()
-    };
-
-    logger.info('Starting IoT Agent listening on port [%s]', config.server.port);
-    logger.debug('Using config:\n\n%s\n', JSON.stringify(config, null, 4));
-
-    contextServer.app.set('port', config.server.port);
-    contextServer.app.set('host', '0.0.0.0');
-    contextServer.app.use(express.json());
-    contextServer.app.use(express.urlencoded());
-
-    if (config.logLevel && config.logLevel === 'DEBUG') {
-        contextServer.app.use(traceRequest);
-    }
-
-    contextServer.app.post('/NGSI10/updateContext', handleUpdate);
-    contextServer.app.post('/NGSI10/queryContext', handleQuery);
-    contextServer.app.post('/notification', handleNotification);
-
-    contextServer.app.use(handleError);
-
-    contextServer.server = http.createServer(contextServer.app);
-
-    contextServer.server.listen(contextServer.app.get('port'), contextServer.app.get('host'), callback);
+function loadContextRoutes(app) {
+    app.post('/NGSI10/updateContext', handleUpdate);
+    app.post('/NGSI10/queryContext', handleQuery);
+    app.post('/notification', handleNotification);
 }
 
-function stop(callback) {
-    logger.info('Stopping IoT Agent');
-
-    contextServer.server.close(callback);
-}
 
 function setUpdateHandler(newHandler) {
     updateHandler = newHandler;
@@ -264,8 +219,6 @@ function setQueryHandler(newHandler) {
     queryHandler = newHandler;
 }
 
+exports.loadContextRoutes = loadContextRoutes;
 exports.setUpdateHandler = setUpdateHandler;
 exports.setQueryHandler = setQueryHandler;
-exports.start = start;
-exports.stop = stop;
-
